requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiofiles=23.2.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
aom=3.6.0=h6a678d5_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h5eee18b_9
brotli-bin=1.0.9=h5eee18b_9
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.4.26=pypi_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=pypi_0
clip=1.0=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
dav1d=1.2.1=h5eee18b_0
exceptiongroup=1.3.0=pypi_0
fastapi=0.115.12=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.58.0=pypi_0
freetype=2.13.3=h4a9f257_0
fsspec=2025.3.2=pypi_0
ftfy=6.3.1=pypi_0
gensim=4.3.3=pypi_0
giflib=5.2.2=h5eee18b_0
googletrans=4.0.2=pypi_0
gradio=4.44.1=pypi_0
gradio-client=1.3.0=pypi_0
h11=0.16.0=pypi_0
h2=4.2.0=pypi_0
hf-xet=1.1.0=pypi_0
hpack=4.1.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.31.1=pypi_0
hyperframe=6.1.0=pypi_0
idna=3.10=pypi_0
imagecodecs=2024.9.22=py39h652d938_1
imageio=2.37.0=py39h06a4308_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=pypi_0
joblib=1.5.0=pypi_0
jpeg=9e=h5eee18b_3
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.7=pypi_0
krb5=1.20.1=h143b758_1
lazy_loader=0.4=py39h06a4308_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libaec=1.1.3=h6a678d5_0
libavif=1.1.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libcurl=8.12.1=hc9e6f67_0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.7.0=hde9077f_0
libwebp-base=1.3.2=h5eee18b_1
libzopfli=1.0.3=he6710b0_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.4=pypi_0
mdurl=0.1.2=pypi_0
medclip=0.0.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_2
mkl_fft=1.3.11=py39h5eee18b_0
mkl_random=1.2.8=py39h1128e8f_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=py39h06a4308_0
nltk=3.9.1=pypi_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-cublas-cu12=12.6.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.6.80=pypi_0
nvidia-cuda-nvrtc-cu12=12.6.77=pypi_0
nvidia-cuda-runtime-cu12=12.6.77=pypi_0
nvidia-cudnn-cu12=9.5.1.17=pypi_0
nvidia-cufft-cu12=11.3.0.4=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.7.77=pypi_0
nvidia-cusolver-cu12=11.7.1.2=pypi_0
nvidia-cusparse-cu12=12.5.4.2=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.6.77=pypi_0
opencv-python=4.11.0.86=pypi_0
openjpeg=2.5.2=h0d4d230_1
openssl=3.0.16=h5eee18b_0
orjson=3.10.18=pypi_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pillow=10.4.0=pypi_0
pip=25.1=pyhc872135_2
pybind11-abi=4=hd3eb1b0_1
pydantic=2.11.5=pypi_0
pydantic-core=2.33.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.3=pypi_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.17=pypi_0
pyqt5-sip=12.17.0=pypi_0
python=3.9.21=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=14.0.0=pypi_0
ruff=0.11.12=pypi_0
safetensors=0.5.3=pypi_0
scikit-image=0.24.0=py39h1128e8f_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=py39h5f9d8c6_1
semantic-version=2.10.0=pypi_0
setuptools=78.1.1=py39h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smart-open=7.1.0=pypi_0
snappy=1.2.1=h6a678d5_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.46.2=pypi_0
sympy=1.14.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
textaugment=2.0.0=pypi_0
textblob=0.19.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2023.4.12=py39h06a4308_0
timm=1.0.15=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.13.3=pypi_0
tomlkit=0.12.0=pypi_0
torch=2.7.0=pypi_0
torchvision=0.22.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.24.0=pypi_0
triton=3.3.0=pypi_0
typer=0.16.0=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pypi_0
uvicorn=0.34.3=pypi_0
wcwidth=0.2.13=pypi_0
websockets=12.0=pypi_0
wget=3.2=pypi_0
wheel=0.45.1=py39h06a4308_0
wrapt=1.17.2=pypi_0
xz=5.6.4=h5eee18b_1
zfp=1.0.0=h6a678d5_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
zlib-ng=2.0.7=h5eee18b_0
zstd=1.5.6=hc292b87_0
